+++ py_abac/storage/memory/__init__.py
@@ -1,3 +1,4 @@
 """
     In-Memory storage
 """
+from .storage import MemoryStorage